Add basic support for IPv6
Add functions parse-ipv6 and ipv6-string for basic conversion between
string and integer representations of IPv6 addresses.

parse-ipv6 converts an IPv6 string into an integer, and handles the
usual variations in formatting.

parse-ipv6-cidr interprets a string such as "cafe:beef::/64" and returns
multiple values: the integer version of the address, per parse-ipv6, and
the prefix-length. Both return values are integers.

ipv6-string converts a string into the strictest (and simplest)
full-length representation - it does not remove leading zeroes, or
replace sequences of 2+ zero-sections with a double-colon.

Also add basic tests.

Update the README to mention the new stuff.